backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]

# Core Dependencies
python-dotenv
pydantic
jinja2
aiofiles
requests

# Google Gemini
google-generativeai==0.3.0

# CrewAI & LangChain (TESTED COMPATIBLE VERSIONS)
crewai==0.28.8
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0